backend/pyproject.toml
[project]
name = "video-manual-generator"
version = "0.1.0"
description = "Video から自動マニュアル生成システム"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "opencv-python>=4.8.1",
    "numpy>=1.26.0",
    "Pillow>=10.1.0",
    "openai-whisper>=20231117",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
all = [
    "video-manual-generator[dev,test,export]",
]
dev = [
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.1",
]
export = [
    "playwright>=1.40.0",
    "weasyprint>=60.1",
    "markdown>=3.5.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
